update(test): sep test_funcsions.py to test_function/
