+++ test/integration_test/run_original.py
@@ -1,4 +1,4 @@
-from test_functions import test
+from test_function.test_functions import test
 
 
 def main():
